fixed build process, refactored Util into seperate classes to comply with SoC
