+++ index.html
@@ -4,11 +4,9 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Demo Class</title>
+    <title>UA Engine</title>
   </head>
   <body>
-    <script src="./libs/SpinePlugin.min.js"></script>
-    <script src="./libs/phaser.min.js"></script>
     <script src="./libs/main.js"></script>
     <script>
      /* System.import('./libs/main.js').then(function(m){
@@ -16,7 +14,7 @@
       }); */
      // let m = new window.System.main(); 
     </script>
-    <div id="phaser-game"></div>
+    <div id="ua-engine"></div>
     <style>
       body,
       html {
