+++ src/scenes/types/globals.d.ts
@@ -1,0 +1,6 @@
+// https://www.typescriptlang.org/docs/handbook/declaration-merging.html
+declare interface Window
+{
+	SpinePlugin: any
+}
+
