+++ typings/phaser3-spine-plugin.d.ts
