+++ typings/spine.d.ts
