+++ typings/spine.plugin.d.ts
